+++ lua/community.lua
@@ -97,4 +97,19 @@ return {
   -- ----------------------------------------------
 
   -- ----------------------------------------------
+  -- Utility
+  -- rich command prompt
+  { import = "astrocommunity.utility.noice-nvim" },
+  -- Override notification display defaults
+  {
+    "rcarriga/nvim-notify",
+    opts = {
+      top_down = false,
+      timeout = 1000, -- shorter display duration, default 1000
+      -- log messages level - default 5 (everything), 1 (minimum)
+      level = 3,
+      -- background_color = "#000000",
+    },
+  },
+  -- ----------------------------------------------
 }